requirements.txt
certifi >= 2021.5.30
cffi >= 1.14.6
charset-normalizer >= 2.0.4
configparser >= 5.0.2
cryptography >= 3.4.7
idna >= 3.2
netaddr >= 0.8.0
numpy >= 1.21.1
pycparser >= 2.20
pyOpenSSL >= 20.0.1
requests >= 2.26.0
RouterOS-api >= 0.17.0
six >= 1.16.0
urllib3 >= 1.26.6
dnspython >= 2.1.0